Create DirectoryInfoExtensions
